+++ src/btConvexHull/btConvexHullComputer.cpp
@@ -1767,8 +1767,8 @@ static bool pointCmp(const btConvexHullInternal::Point32& p, const btConvexHullI
 void btConvexHullInternal::compute(vector<array<double, 3>> vertices)
 {
     btVector3 min(btScalar(1e30), btScalar(1e30), btScalar(1e30)), max(btScalar(-1e30), btScalar(-1e30), btScalar(-1e30));
-    size_t count = (int)vertices.size();
-    for (int32_t i = 0; i < count; i++) {
+    size_t count = vertices.size();
+    for (size_t i = 0; i < count; i++) {
         btVector3 p((btScalar)vertices[i][0], (btScalar)vertices[i][1], (btScalar)vertices[i][2]);
         min.setMin(p);
         max.setMax(p);
@@ -1802,7 +1802,7 @@ void btConvexHullInternal::compute(vector<array<double, 3>> vertices)
 
     btAlignedObjectArray<Point32> points;
     points.resize(count);
-    for (int32_t i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         btVector3 p((btScalar)vertices[i][0], (btScalar)vertices[i][1], (btScalar)vertices[i][2]);
         p = (p - center) * s;
         points[i].x = (int32_t)p[medAxis];
@@ -1815,7 +1815,7 @@ void btConvexHullInternal::compute(vector<array<double, 3>> vertices)
     vertexPool.reset();
     vertexPool.setArraySize(count);
     originalVertices.resize(count);
-    for (int32_t i = 0; i < count; i++) {
+    for (size_t i = 0; i < count; i++) {
         Vertex* v = vertexPool.newObject();
         v->edges = NULL;
         v->point = points[i];
